+++ src/App.js
@@ -1,37 +1,23 @@
-import React, { useState, useRef } from 'react';
+import React, { useState, useCallback } from 'react';
 
 import Tree from './components/Tree/Tree';
 import data from './components/Tree/data';
 
 import './App.scss';
-
-function updateTree(tree, value, newValue, key = 'id', reverse = false) {
-  const stack = [tree];
-  while (stack.length) {
-    const node = stack[reverse ? 'pop' : 'shift']();
-    if (node[key] === value) {
-      node.value = newValue;
-      return tree;
-    }
-    node.children && stack.push(...node.children);
-  }
-  return null;
-}
+import { updateTree } from './utils';
 
 function App() {
-  const textInput = useRef(null);
   const [nodeEdit, setNodeEdit] = useState('');
   const [nodeName, setNodeName] = useState('');
   const [dataSource, setDataSource] = useState(data);
 
   const handleSetIdEditNode = (text) => {
     setNodeEdit(text);
-    textInput.current.focus();
   };
 
-  const handleChangeInput = (event) => {
+  const handleChangeInput = useCallback((event) => {
     setNodeName(event.target.value);
-  };
+  }, []);
 
   const handleSave = () => {
     let tree = {};
@@ -51,7 +37,6 @@ function App() {
       <div className="input-zone">
         <label>Name</label>
         <input
-          ref={textInput}
           value={nodeName}
           onChange={handleChangeInput}
           placeholder="Enter node name"
